// src/types/yandex/notification-types.ts
export interface GoodsFeedbackCreatedNotificationDTO {
    /**
     * Тип уведомления:
     *
     * * `PING` — проверочное уведомление.
     * * `ORDER_CREATED` — создан новый заказ.
     * * `ORDER_CANCELLED` — заказ отменен.
     * * `ORDER_STATUS_UPDATED` — статус заказа изменен.
     * * `ORDER_RETURN_CREATED` — создан новый невыкуп или возврат.
     * * `ORDER_CANCELLATION_REQUEST` — создана заявка на отмену заказа (для DBS-магазинов).
     * * `ORDER_RETURN_STATUS_UPDATED` — статус невыкупа или возврата изменен.
     * * `GOODS_FEEDBACK_CREATED` — создан новый отзыв о товаре.
     * * `GOODS_FEEDBACK_COMMENT_CREATED` — создан новый комментарий к отзыву о товаре.
     * * `CHAT_CREATED` — создан новый чат с покупателем.
     * * `CHAT_MESSAGE_SENT` — добавлено новое сообщение в чате.
     * * `CHAT_ARBITRAGE_STARTED` — по обращению покупателя начался спор.
     * * `CHAT_ARBITRAGE_FINISHED` — спор завершен.
     */
    notificationType: NotificationType
    /**
     * Идентификатор отзыва.
     * @format int64
     */
    feedbackId: number
    /**
     * Идентификатор кабинета.
     * @format int64
     */
    businessId: number
    /**
     * Дата и время создания отзыва.
     *
     * Может отличаться от информации в `publishedAt`, так как некоторое время отзыв проходит модерацию.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    createdAt: string
    /**
     * Дата и время публикации отзыва.
     *
     * Может отличаться от информации в `createdAt`, так как некоторое время отзыв проходит модерацию.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    publishedAt: string
}
// ...
export enum NotificationType {
    PING = 'PING',
    ORDER_CREATED = 'ORDER_CREATED',
    ORDER_CANCELLED = 'ORDER_CANCELLED',
    ORDER_STATUS_UPDATED = 'ORDER_STATUS_UPDATED',
    ORDER_RETURN_CREATED = 'ORDER_RETURN_CREATED',
    ORDER_CANCELLATION_REQUEST = 'ORDER_CANCELLATION_REQUEST',
    ORDER_RETURN_STATUS_UPDATED = 'ORDER_RETURN_STATUS_UPDATED',
    GOODS_FEEDBACK_CREATED = 'GOODS_FEEDBACK_CREATED',
    GOODS_FEEDBACK_COMMENT_CREATED = 'GOODS_FEEDBACK_COMMENT_CREATED',
    CHAT_CREATED = 'CHAT_CREATED',
    CHAT_MESSAGE_SENT = 'CHAT_MESSAGE_SENT',
    CHAT_ARBITRAGE_STARTED = 'CHAT_ARBITRAGE_STARTED',
    CHAT_ARBITRAGE_FINISHED = 'CHAT_ARBITRAGE_FINISHED'
}